Refactor e2e
- Refactor e2e
